+++ src/index/arctic.rs
@@ -1,12 +1,12 @@
 use crate::Index;
 use crate::index;
 
-impl<K, H> Index<K, H> for arctic::concurrent::Map<K, u32>
+impl<K, H> Index<K, H> for arctic::concurrent::Map<K, u64>
 where
     K: index::Key,
     H: index::Hasher,
 {
-    type Send<'a> = &'a arctic::concurrent::Map<K, u32>;
+    type Send<'a> = &'a arctic::concurrent::Map<K, u64>;
 
     fn new(config: &index::Config) -> Self {
         arctic::concurrent::Map::with_reclaim_threshold(config.reclaim_threshold)
@@ -22,13 +22,13 @@ where
     }
 }
 
-impl<K, H> index::IndexSend<K, H> for &'_ arctic::concurrent::Map<K, u32>
+impl<K, H> index::IndexSend<K, H> for &'_ arctic::concurrent::Map<K, u64>
 where
     K: index::Key,
     H: index::Hasher,
 {
     type Handle<'a>
-        = arctic::concurrent::MapRef<'a, K, u32>
+        = arctic::concurrent::MapRef<'a, K, u64>
     where
         Self: 'a;
 
@@ -37,27 +37,27 @@ where
     }
 }
 
-impl<'a, K> index::IndexPin<K> for arctic::concurrent::MapRef<'a, K, u32>
+impl<'a, K> index::IndexPin<K> for arctic::concurrent::MapRef<'a, K, u64>
 where
     K: index::Key,
 {
-    fn get(&mut self, key: &K) -> Option<u32> {
+    fn get(&mut self, key: &K) -> Option<u64> {
         arctic::concurrent::MapRef::get(self, key.borrow())
     }
 
-    fn insert(&mut self, key: K, value: u32) -> Option<u32> {
+    fn insert(&mut self, key: K, value: u64) -> Option<u64> {
         arctic::concurrent::MapRef::upsert(self, key.borrow(), value)
     }
 
-    fn update(&mut self, key: K, value: u32) -> Option<u32> {
+    fn update(&mut self, key: K, value: u64) -> Option<u64> {
         arctic::concurrent::MapRef::update(self, key.borrow(), value).ok()
     }
 
-    fn increment(&mut self, key: K) -> Option<u32> {
+    fn increment(&mut self, key: K) -> Option<u64> {
         arctic::concurrent::MapRef::upsert_with(self, key.borrow(), |old| old.unwrap_or(0) + 1)
     }
 
-    fn remove(&mut self, key: K) -> Option<u32> {
+    fn remove(&mut self, key: K) -> Option<u64> {
         arctic::concurrent::MapRef::remove(self, key.borrow())
     }
 
@@ -66,7 +66,7 @@ where
         _retry_scan: usize,
         min: &'b K,
         max: &'b K,
-        output: &mut Vec<(K, u32)>,
+        output: &mut Vec<(K, u64)>,
     ) {
         let Some(guard) = arctic::concurrent::MapRef::range(self, min.borrow(), max.borrow())
         else {
