+++ src/index/papaya.rs
@@ -3,7 +3,7 @@ use core::cell::Cell;
 use crate::Index;
 use crate::index;
 
-impl<K: index::Key, H: index::Hasher> Index<K, H> for papaya::HashMap<K, u32, H> {
+impl<K: index::Key, H: index::Hasher> Index<K, H> for papaya::HashMap<K, u64, H> {
     type Send<'a> = &'a Self;
 
     fn new(_: &index::Config) -> Self {
@@ -15,7 +15,7 @@ impl<K: index::Key, H: index::Hasher> Index<K, H> for papaya::HashMap<K, u32, H>
     }
 }
 
-impl<K: index::Key, H: index::Hasher> index::IndexSend<K, H> for &'_ papaya::HashMap<K, u32, H> {
+impl<K: index::Key, H: index::Hasher> index::IndexSend<K, H> for &'_ papaya::HashMap<K, u64, H> {
     type Handle<'a>
         = Self
     where
@@ -26,28 +26,28 @@ impl<K: index::Key, H: index::Hasher> index::IndexSend<K, H> for &'_ papaya::Has
     }
 }
 
-impl<K: index::Key, H: index::Hasher> index::IndexPin<K> for &'_ papaya::HashMap<K, u32, H> {
-    fn get(&mut self, key: &K) -> Option<u32> {
+impl<K: index::Key, H: index::Hasher> index::IndexPin<K> for &'_ papaya::HashMap<K, u64, H> {
+    fn get(&mut self, key: &K) -> Option<u64> {
         let map = self.pin();
         map.get(key).copied()
     }
 
-    fn insert(&mut self, key: K, value: u32) -> Option<u32> {
+    fn insert(&mut self, key: K, value: u64) -> Option<u64> {
         let map = self.pin();
         map.insert(key, value).copied()
     }
 
-    fn update(&mut self, key: K, value: u32) -> Option<u32> {
+    fn update(&mut self, key: K, value: u64) -> Option<u64> {
         let map = self.pin();
         map.update(key, |_| value).copied()
     }
 
-    fn remove(&mut self, key: K) -> Option<u32> {
+    fn remove(&mut self, key: K) -> Option<u64> {
         let map = self.pin();
         map.remove(&key).copied()
     }
 
-    fn increment(&mut self, key: K) -> Option<u32> {
+    fn increment(&mut self, key: K) -> Option<u64> {
         let map = self.pin();
         let old = Cell::new(None);
         map.update_or_insert(
